refactor(web): extract ReferralContent so it can render in a panel or modal
